perf(dashboard): Improve perf of highlighting charts in scope of active filter
